Code Dump
One more mass-upload! Rescued this from my old DropBox, and it's dated June 2015, which is cool.

(15 year old me made 4 projects in 1 year. 19 year old me can't finish 1 hahaha)

Yes, so this is Bit Hunter but with more content in every field. It's tied in to a Kung Fu teacher's business, with themed images and audio playing and appearing throughout.

The score system is much more interesting, with multipliers, combos and streaks all in this build.

Probably the best game I had made before I started again in 2019.